refactor preproc.py into two separate files
