+++ brainglobe_template_builder/preproc/__init__.py
@@ -1,0 +1,6 @@
+from brainglobe_template_builder.preproc.masking import create_mask
+from brainglobe_template_builder.preproc.alignment import (
+    get_midline_points,
+    get_alignment_transform,
+    apply_transform,
+)
